% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saebenchmarking-package.R
\docType{package}
\name{saebenchmarking-package}
\alias{saebenchmarking}
\alias{saebenchmarking-package}
\title{saebenchmarking: Benchmarking Small Area Estimates and Their Mean Squared Errors}
\description{
Adjusts model-based small area estimates so that their weighted aggregate agrees with the weighted aggregate of the direct estimates, using the difference, ratio, and optimum benchmarking methods described in Rao and Molina (2015, ISBN:978-1-118-73578-7) and Wang, Fuller and Qu (2008). The mean squared error (MSE) of the benchmarked empirical best linear unbiased predictor (EBLUP) under the Fay-Herriot model is estimated with the second-order approximation or the parametric bootstrap of Steorts and Ghosh (2013) \doi{10.5705/ss.2012.053}. The posterior MSE of the benchmarked hierarchical Bayes (HB) estimator follows Datta, Ghosh, Steorts and Maples (2011) \doi{10.1007/s11749-010-0218-y}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fionaaudia/saebenchmarking}
  \item Report bugs at \url{https://github.com/fionaaudia/saebenchmarking/issues}
}

}
\author{
\strong{Maintainer}: Fiona Audia Nauli Sihombing \email{fionasihombing95@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Fiona Audia Nauli Sihombing \email{fionasihombing95@gmail.com} [copyright holder]
  \item Azka Ubaidillah
}

}
\keyword{internal}
